add "garbage collector" to ProcManager
If a process exits of its own accord, the ProcManager needs to remove it
from its list of running processes. Add a method called at initialization
that spawns a thread to run through the list of running processes and
remove them from the list if they've exited. The loop in the thread
runs once a milisecond.