attempt to improve vlc killing
VLC is misbehaving. Sometimes, it exits on keypress like it
should, but other times it just sticks around until the video
ends.

Try spawning a thread that kills VLC after a few seconds, but
decide against it because it turns out to be messy. Keep code in
a comment just in case I need it while the gallery show is going
on. Comment out the 'clear' line too since right now it's removing
the line of text output after the video exits.

In the end, decide to just shorten the video file to a few
seconds. It's not perfect but it'll have to do under the time
constraints.